+++ govcd/edgegateway_test.go
@@ -7,6 +7,9 @@
 package govcd
 
 import (
+	"io/ioutil"
+	"net/http"
+	"regexp"
 	"strings"
 
 	"github.com/vmware/go-vcloud-director/v2/types/v56"
@@ -537,4 +540,96 @@ func (vcd *TestVCD) Test_UpdateNATRule(check *C) {
 	check.Assert(err, IsNil)
 
 	check.Assert(len(edge.EdgeGateway.Configuration.EdgeGatewayServiceConfiguration.NatService.NatRule), Equals, beforeChangeNatRulesNumber)
+}
+
+// TestEdgeGateway_UpdateLBGeneralParams main point is to test that no load balancer configuration
+// xml tags are lost during changes of load balancer main settings (enable, logging)
+// The test does following steps:
+// 1. Cache raw XML body and marshaled struct in variables before running the test
+// 2. Toggle the settings of load balancer in various ways and ensure no err is returned
+// 3. Set the settings back as they originally were and again get raw XML body and marshaled struct
+// 4. Compare the XML text and structs before configuration and after configuration - they should be
+// identical except <version></version> tag which is versioning the configuration
+func (vcd *TestVCD) TestEdgeGateway_UpdateLBGeneralParams(check *C) {
+	if vcd.config.VCD.EdgeGateway == "" {
+		check.Skip("Skipping test because no edge gatway given")
+	}
+	edge, err := vcd.vdc.FindEdgeGateway(vcd.config.VCD.EdgeGateway)
+	check.Assert(err, IsNil)
+
+	if !edge.HasAdvancedNetworking() {
+		check.Skip("Skipping test because the edge gateway does not have advanced networking enabled")
+	}
+
+	// Cache current load balancer settings for change validation in the end
+	beforeLb, beforeLbXml := testCacheLoadBalancer(edge, check)
+
+	_, err = edge.UpdateLBGeneralParams(true, true, true, "critical")
+	check.Assert(err, IsNil)
+
+	_, err = edge.UpdateLBGeneralParams(false, false, false, "emergency")
+	check.Assert(err, IsNil)
+
+	// Try to set invalid loglevel to get validation error
+	_, err = edge.UpdateLBGeneralParams(false, false, false, "invalid_loglevel")
+	check.Assert(err, ErrorMatches, ".*Valid log levels are.*")
+
+	// Restore to initial settings and validate that it
+	_, err = edge.UpdateLBGeneralParams(beforeLb.Enabled, beforeLb.AccelerationEnabled,
+		beforeLb.Logging.Enable, beforeLb.Logging.LogLevel)
+	check.Assert(err, IsNil)
+
+	// Validate load balancer configuration against initially cached version
+	testCheckLoadBalancerConfig(beforeLb, beforeLbXml, edge, check)
+}
+
+// testGetLBGeneralParamsXML is used for additional validation that modifying load balancer
+// does not change any single field. It returns a string of whole load balancer configuration
+func testGetLBGeneralParamsXML(edge EdgeGateway, check *C) string {
+
+	httpPath, err := edge.buildProxiedEdgeEndpointURL(types.LBConfigPath)
+	check.Assert(err, IsNil)
+
+	resp, err := edge.client.ExecuteRequestWithCustomError(httpPath, http.MethodGet, types.AnyXMLMime,
+		"unable to get XML from load balancer %s", nil, &types.NSXError{})
+	check.Assert(err, IsNil)
+
+	defer resp.Body.Close()
+
+	body, err := ioutil.ReadAll(resp.Body)
+	check.Assert(err, IsNil)
+
+	return string(body)
+}
+
+// cacheLoadBalancer is meant to store load balancer settings before any operations so that all
+// configuration can be checked after manipulation
+func testCacheLoadBalancer(edge EdgeGateway, check *C) (*types.LBGeneralParamsWithXML, string) {
+	beforeLb, err := edge.GetLBGeneralParams()
+	check.Assert(err, IsNil)
+	beforeLbXml := testGetLBGeneralParamsXML(edge, check)
+	return beforeLb, beforeLbXml
+}
+
+// testCheckLoadBalancerConfig validates if both raw XML string and load balancer struct remain
+// identical after settings manipulation.
+func testCheckLoadBalancerConfig(beforeLb *types.LBGeneralParamsWithXML, beforeLbXml string, edge EdgeGateway, check *C) {
+	afterLb, err := edge.GetLBGeneralParams()
+	check.Assert(err, IsNil)
+
+	afterLbXml := testGetLBGeneralParamsXML(edge, check)
+
+	// remove `<version></version>` tag from both XML represntation and struct for deep comparison
+	// because this version changes with each update and will never be the same after a few
+	// operations
+
+	reVersion := regexp.MustCompile(`<version>\w*<\/version>`)
+	beforeLbXml = reVersion.ReplaceAllLiteralString(beforeLbXml, "")
+	afterLbXml = reVersion.ReplaceAllLiteralString(afterLbXml, "")
+
+	beforeLb.Version = ""
+	afterLb.Version = ""
+
+	check.Assert(beforeLb, DeepEquals, afterLb)
+	check.Assert(beforeLbXml, DeepEquals, afterLbXml)
 }