+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.3.0 (Unreleased)
 
 * Added edge gateway create/delete functions [#130](https://github.com/vmware/go-vcloud-director/issues/130).
+* Added edge gateway global load balancer configuration support (e.g. enable/disable) [#219](https://github.com/vmware/go-vcloud-director/pull/219)
 * Added load balancer service monitor [#196](https://github.com/vmware/go-vcloud-director/pull/196)
 * Added load balancer server pool [#205](https://github.com/vmware/go-vcloud-director/pull/205)
 * Added load balancer application profile [#208](https://github.com/vmware/go-vcloud-director/pull/208)